[refactor] rename to appbar
